+++ react/components/AccordionFilterGroup.js
@@ -21,6 +21,9 @@ const AccordionFilterGroup = ({
   onOpen,
   onFilterCheck,
   appliedFiltersOverview,
+  navigationType,
+  initiallyCollapsed,
+  truncateFilters,
 }) => {
   const handles = useCssHandles(CSS_HANDLES)
   const filters = useSelectedFilters(facets)
@@ -37,6 +40,8 @@ const AccordionFilterGroup = ({
       onOpen={onOpen}
       selectedFilters={selectedFilters}
       appliedFiltersOverview={appliedFiltersOverview}
+      navigationType={navigationType}
+      initiallyCollapsed={initiallyCollapsed}
     >
       <div
         className={classNames(
@@ -48,6 +53,8 @@ const AccordionFilterGroup = ({
           onFilterCheck={onFilterCheck}
           facets={filters}
           facetTitle={facetTitle}
+          truncateFilters={truncateFilters}
+          navigationType={navigationType}
         />
       </div>
     </AccordionFilterItem>