+++ react/components/FacetCheckboxList.js
@@ -7,13 +7,22 @@ import styles from '../searchResult.css'
 import SettingsContext from '../components/SettingsContext'
 import { searchSlugify } from '../utils/slug'
 import { SearchFilterBar } from './SearchFilterBar'
+import { FACETS_RENDER_THRESHOLD } from '../constants/filterConstants'
+import ShowMoreFilterButton from './ShowMoreFilterButton'
 
 const useSettings = () => useContext(SettingsContext)
 
-const FacetCheckboxList = ({ facets, onFilterCheck, facetTitle }) => {
+const FacetCheckboxList = ({
+  facets,
+  onFilterCheck,
+  facetTitle,
+  truncateFilters,
+  navigationType,
+}) => {
   const { showFacetQuantity } = useContext(SettingsContext)
   const { thresholdForFacetSearch } = useSettings()
   const [searchTerm, setSearchTerm] = useState('')
+  const [truncated, setTruncated] = useState(true)
 
   const filteredFacets = useMemo(() => {
     if (thresholdForFacetSearch === undefined || searchTerm === '') {
@@ -24,6 +33,17 @@ const FacetCheckboxList = ({ facets, onFilterCheck, facetTitle }) => {
     )
   }, [facets, searchTerm, thresholdForFacetSearch])
 
+  const shouldTruncate =
+    navigationType === 'collapsible' &&
+    truncateFilters &&
+    // The "+ 1" prevents from truncating a single value 
+    filteredFacets.length > FACETS_RENDER_THRESHOLD + 1
+
+  const endSlice =
+    shouldTruncate && truncated
+      ? FACETS_RENDER_THRESHOLD
+      : filteredFacets.length
+
   const showSearchBar =
     thresholdForFacetSearch !== undefined &&
     thresholdForFacetSearch < facets.length
@@ -33,7 +53,7 @@ const FacetCheckboxList = ({ facets, onFilterCheck, facetTitle }) => {
       {showSearchBar ? (
         <SearchFilterBar name={facetTitle} handleChange={setSearchTerm} />
       ) : null}
-      {filteredFacets.map(facet => {
+      {filteredFacets.slice(0, endSlice).map(facet => {
         const { name } = facet
         const slugifiedName = searchSlugify(name)
 
@@ -62,6 +82,13 @@ const FacetCheckboxList = ({ facets, onFilterCheck, facetTitle }) => {
           </div>
         )
       })}
+      {shouldTruncate && (
+        <ShowMoreFilterButton
+          quantity={facets.length - FACETS_RENDER_THRESHOLD}
+          truncated={truncated}
+          toggleTruncate={() => setTruncated(!truncated)}
+        />
+      )}
     </>
   )
 }